Rotation&Zoom combination error resolved

+++ cursus/core/fdf/src/for_each_point.c
@@ -6,7 +6,7 @@
 /*   By: fcorri <fcorri@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/07/25 16:15:13 by fcorri            #+#    #+#             */
-/*   Updated: 2023/08/02 17:01:46 by fcorri           ###   ########.fr       */
+/*   Updated: 2023/08/03 13:20:19 by fcorri           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -29,7 +29,7 @@ void	ft_vvv_for_each_point_of(t_vars *vars, t_v3 (*op)(t_v3 a, t_v3 b), \
 	}
 }
 
-void	ft_vvs_for_each_point_of(t_vars *vars, t_v3 (*op)(t_v3 a, int k), int k)
+void	ft_vvs_for_each_point_of(t_vars *vars, t_v3 (*op)(t_v3 a, float k), float k)
 {
 	int	row;
 	int	col;
